[crashtracker] Refactor receiver code into seperate files.
